Remove unnecessary GradingDashboard abstraction
